requirements.txt
langchain
langchain-openai
langchain-community
langgraph
faiss-cpu
pypdf
python-dotenv
streamlit
tiktoken
reportlab